Add content update listener (from CodeMirror)
CodeMirror allows you to add extensions like basicSetup inside the
extensions array. This addition (EditorView.updateListener.of() method)
creates a listener that runs on editor events like input, paste, etc.
When the document changes (update.docChanged), it calls the private
`#update` method.

[updateListener]
https://codemirror.net/docs/ref/#view.EditorView^updateListener
(the .of() method must be defined somewhere in the CodeMirror codebase)