[LEARNING] If we weren't using fat arrow funcs...
The fat arrow function was introduced in ES6 and is similar to Ruby's
`->(lambda)` syntax - but it has a lexical `this` binding.

This means that it preserves the value of `this` from the surrounding
context. Before ES6, you had to bind it using `bind(this)`.

If we were to remove the `bind(this)`, we'd get a bunch of errors.
Without `bind(this)` or that fat arrow function, `this` refers to the
CodeMirror editor's context/scope-not the Stimulus controller. So
calling `this.#update` fails because #update is defined in the
editor_controller, not in CodeMirror.